Break out member team memberships into a component
